requirements.txt
streamlit>=1.32.0
reportlab>=4.0.0
Pillow>=10.0.0
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
rembg>=2.0.50
onnxruntime>=1.17.0
numpy>=1.24.0
scipy>=1.11.0
